adds address checker, issues with sdk

+++ src/index.js
@@ -1,4 +1,4 @@
-import { Client, DMChannel, Intents, Interaction } from 'discord.js';
+import { Client, DMChannel, Intents, Interaction, Message } from 'discord.js';
 import { ethers } from 'ethers';
 import { promisify } from 'util';
 import { SDK } from '@elastic-dao/sdk';
@@ -27,26 +27,20 @@ const sdkConfig = {
   storageAdapter: new RedisAdapter(),
 };
 
+console.log(sdkConfig);
+
+const client = new Client({
+  intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.DIRECT_MESSAGES],
+});
+
 const sdk = new SDK(sdkConfig);
 const redis = Redis.createClient();
 const getAsync = promisify(redis.get).bind(redis);
 const setAsync = promisify(redis.set).bind(redis);
-const isAddress = ethers.utils.isAddress().toString();
-
-const detectNewETHAddress = async (str) => {
-  const address = detectETHAddress(str);
-
-  if (address) {
-    if (!(await getAsync(address))) {
-      return address;
-    }
-  }
-
-  return false;
-};
-
-// Create a new client instance
-const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES] });
+const isAddress = (str) => ethers.utils.isAddress(str);
+const dao = await sdk.models.DAO.deserialize('0xaaa1f5fc9617195b5aa7fd1bd989d47f9e8d3f82');
+console.log(dao.toObject());
+const token = await dao.token();
 
 const registerAddress = async (user, address) => {
   const EGT_HOLDER = new Set(JSON.parse((await getAsync('EGTHolder')) || '[]'));
@@ -67,47 +61,48 @@ class Ready {
     const { author, channel, content } = messageCreate;
     const isDM = messageCreate.channel.type == 'DM';
     const dmChannel = isDM ? channel : await author.createDM();
-    console.log('Got a generic message', content, 'on channel', channel.id, 'from user', author);
+    console.log(
+      `Got a generic message, ${content}, on channel, ${channel.id}, from user, ${author}`
+    );
+
+    if (author.bot) return;
 
     if (messageCreate.content.startsWith('!')) {
       messageCreate.delete();
-      this.guildMemberJoin(messageCreate);
+      this.guildMemberJoin(messageCreate, dmChannel);
       return;
+    }
+
+    const address = await content.toString();
+
+    const checkedAddress = await isAddress(address);
+
+    console.log('DAO', dao.ecosystem);
+
+    const checkingEGT = await sdk.models.TokenHolder.deserialize(address, dao.ecosystem, token);
+
+    if (checkedAddress == true) {
+      console.log('THIS IS AN ADDRESS');
+    }
+
+    if (checkingEGT && isDM) {
+      if (checkingEGT.lambda > 0) {
+        console.log(checkingEGT.lambda);
+        dmChannel.send('you got some EGT');
+      }
+      dmChannel.send(`You don't have EGT`);
     }
   }
 
-  async guildMemberJoin({ author }) {
-    const dmChannel = await author.createDM();
+  async guildMemberJoin({ author }, dmChannel) {
     const text = await dmChannel.send(
       `Hey Hey! I've heard that you want to verify your address as an EGT Holder.`
     );
+    const filter = () => true;
 
-    const filter = (m) => m.content.includes('test');
-    const collector = dmChannel.createMessageCollector(filter, { time: 15000 });
-
-    collector.on('collect', (m) => console.log('collected item', m.content));
-    collector.on('end', (collected) => {
-      console.log('collected Items', collected.size);
-    });
-
-    // const filter = (m) => m.content.includes('test');
-
-    // const collected = await dmChannel.awaitMessages({ filter, time: 30000 });
-
-    // const collector = dmChannel.createMessageCollector({ filter, time: 30000 });
-
-    // dmChannel
-    //   .awaitMessages({ filter, max: 1, time: 30000, errors: ['time'] })
-    //   .then((collected) => console.log('The filter is working', collected.size))
-    //   .catch((collected) =>
-    //     console.log(`The time for collecting your address has expired, please repeat the process`)
-    //   );
-
-    await text;
-    // this is how to get a username
-    console.log(author.username);
-
-    //await dmChannel.send('please provide your Address on which you have EGT');
+    dmChannel
+      .awaitMessages({ filter, max: 1, time: 5000 })
+      .then((messages) => console.log('messages', messages));
   }
 }
